[FEATURE] Add Controller, Service for Purchase and Swagger documentation

+++ src/main/java/com/projectjavaspring/marketapi/persistence/entity/Cliente.java
@@ -19,7 +19,7 @@ public class Cliente {
 
     private String surnames;
 
-    private Double cellphone;
+    private Long cellphone;
 
     private String address;
